added chat implementation
can create a chat with only the user who created the chat. PUT for adding members to a chat is set up but i cant fiqure out how to add a different member to a chat. DELETE for removing the current user from a chat(deleting it) should work but userInfoViewModel.getEmail() is getting null rn